# packages/frontend/src/redux/schema/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions/{id}": {
      "get": {
        "operationId": "Transactions_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetailsDto"
                }
              }
            }
          }
        },
        "tags": ["Transactions"]
      }
    },
    "/api/transaction-categories": {
      "post": {
        "operationId": "TransactionCategories_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "properties": { "payload": { "type": "string" } } }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "get": {
        "operationId": "TransactionCategories_findAllByUser",
        "parameters": [
          {
            "name": "visibilityType",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/visibilityType" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction category by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionCategoryDetailsDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      }
    },
    "/api/transaction-categories/{id}": {
      "get": {
        "operationId": "TransactionCategories_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCategoryDetailsDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "patch": {
        "operationId": "TransactionCategories_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCategoryDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      },
      "delete": {
        "operationId": "TransactionCategories_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Transaction categories"]
      }
    },
    "/api/transaction-categories/{id}/summary": {
      "get": {
        "operationId": "TransactionCategories_getCategorySummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryMonthlySummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transaction categories"]
      }
    },
...
    "/api/transaction-templates": {
      "post": {
        "operationId": "TransactionTemplates_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "properties": { "payload": { "type": "string" } } }
              }
            }
          }
        },
        "tags": ["Transaction templates"]
      },
      "get": {
        "operationId": "TransactionTemplates_findAllByUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all transaction templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionTemplateDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transaction templates"]
      }
    },
...
    "/api/transaction-templates/{id}": {
      "get": {
        "operationId": "TransactionTemplates_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction template by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionTemplateDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction templates"]
      },
      "patch": {
        "operationId": "TransactionTemplates_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionTemplateDto"
                }
              }
            }
          }
        },
        "tags": ["Transaction templates"]
      },
      "delete": {
        "operationId": "TransactionTemplates_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Transaction templates"]
      }
    },
...
    "/api/transfers": {
      "get": {
        "operationId": "Transfers_findAllByUser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferListItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Transfers"]
      },
      "post": {
        "operationId": "Transfers_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTransferDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransferDetailsDto" }
              }
            }
          }
        },
        "tags": ["Transfers"]
      }
    },
    "/api/transfers/{id}": {
      "get": {
        "operationId": "Transfers_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransferDetailsDto" }
              }
            }
          }
        },
        "tags": ["Transfers"]
      },
      "patch": {
        "operationId": "Transfers_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTransferDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransferDetailsDto" }
              }
            }
          }
        },
        "tags": ["Transfers"]
      },
      "delete": {
        "operationId": "Transfers_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["Transfers"]
      }
    },
...
  },
  "info": {
    "title": "Financer",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TransactionType": {
        "type": "string",
        "enum": ["INCOME", "EXPENSE", "TRANSFER"]
      },
      "TransactionCategoryDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "deleted": { "type": "boolean" },
          "parentCategoryId": { "type": "string", "nullable": true }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "userId",
          "name",
          "visibility",
          "deleted",
          "parentCategoryId"
        ]
      },
...
      "TransactionTemplateType": {
        "type": "string",
        "enum": ["MANUAL", "AUTO"]
      },
      "TransactionTemplateDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "templateName": { "type": "string" },
          "templateType": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionTemplateType" }
          },
          "templateVisibility": {
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "amount": { "type": "number", "nullable": true },
          "description": { "type": "string" },
          "dayOfMonth": { "type": "number", "nullable": true },
          "dayOfMonthToCreate": { "type": "number", "nullable": true },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "categories": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "templateName",
          "templateType",
          "templateVisibility",
          "description",
          "userId"
        ]
      },
...
      "TransactionDetailCategoryDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "amount": { "type": "number" },
          "name": { "type": "string" },
          "path": { "type": "string" }
        },
        "required": ["id", "description", "amount", "name", "path"]
      },
      "TransactionDetailsDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "fromAccountName": { "type": "string", "nullable": true },
          "toAccountName": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" },
          "type": {
            "description": "Type of the transaction",
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDetailCategoryDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "fromAccount",
          "toAccount",
          "fromAccountName",
          "toAccountName",
          "isRecurring",
          "type",
          "categories"
        ]
      },
      "CategoryMonthlySummaryDtoId": {
        "type": "object",
        "properties": {
          "year": { "type": "number" },
          "month": { "type": "number" }
        },
        "required": ["year", "month"]
      },
      "CategoryMonthlySummaryDto": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/CategoryMonthlySummaryDtoId" },
          "totalCount": { "type": "number" },
          "incomesCount": { "type": "number" },
          "expensesCount": { "type": "number" },
          "transfersCount": { "type": "number" },
          "totalAmount": { "type": "number" },
          "incomeAmount": { "type": "number" },
          "expenseAmount": { "type": "number" },
          "transferAmount": { "type": "number" }
        },
        "required": [
          "id",
          "totalCount",
          "incomesCount",
          "expensesCount",
          "transfersCount",
          "totalAmount",
          "incomeAmount",
          "expenseAmount",
          "transferAmount"
        ]
      },
      "CreateTransactionCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "parentCategoryId": { "type": "string", "nullable": true }
        },
        "required": ["name", "visibility", "parentCategoryId"]
      },
      "visibilityType": {
        "type": "string",
        "enum": ["INCOME", "EXPENSE", "TRANSFER"]
      },
      "TransactionCategoryDetailsDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "deleted": { "type": "boolean" },
          "parentCategoryId": { "type": "string", "nullable": true },
          "path": { "type": "string" }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "userId",
          "name",
          "visibility",
          "deleted",
          "parentCategoryId",
          "path"
        ]
      },
      "UpdateTransactionCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "visibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionType" }
          },
          "parentCategoryId": { "type": "string", "nullable": true }
        }
      },
...
      "CreateTransactionTemplateDto": {
        "type": "object",
        "properties": {
          "templateName": { "type": "string" },
          "templateType": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionTemplateType" }
          },
          "templateVisibility": {
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "amount": { "type": "number", "nullable": true },
          "description": { "type": "string" },
          "dayOfMonth": { "type": "number", "nullable": true },
          "dayOfMonthToCreate": { "type": "number", "nullable": true },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "categories": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "templateName",
          "templateType",
          "templateVisibility",
          "description"
        ]
      },
      "UpdateTransactionTemplateDto": {
        "type": "object",
        "properties": {
          "templateName": { "type": "string" },
          "templateType": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionTemplateType" }
          },
          "templateVisibility": {
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "amount": { "type": "number", "nullable": true },
          "description": { "type": "string" },
          "dayOfMonth": { "type": "number", "nullable": true },
          "dayOfMonthToCreate": { "type": "number", "nullable": true },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "categories": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "TransferDetailsDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "fromAccountName": { "type": "string", "nullable": true },
          "toAccountName": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" },
          "type": {
            "description": "Type of the transaction",
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDetailCategoryDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "fromAccount",
          "toAccount",
          "fromAccountName",
          "toAccountName",
          "isRecurring",
          "type",
          "categories"
        ]
      },
      "TransferListItemDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "isRecurring": { "type": "boolean" },
          "type": {
            "description": "Type of the transaction",
            "allOf": [{ "$ref": "#/components/schemas/TransactionType" }]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionListItemCategoryDto"
            }
          }
        },
        "required": [
          "id",
          "amount",
          "description",
          "date",
          "fromAccount",
          "toAccount",
          "isRecurring",
          "type",
          "categories"
        ]
      },
      "CreateTransferDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": ["amount", "description", "date"]
      },
      "UpdateTransferDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "fromAccount": { "type": "string", "nullable": true },
          "toAccount": { "type": "string", "nullable": true },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        }
      }
    }
  }
}